Move read_manifest() logic into Schema
